KAN-36 [REFACTOR] Modify the way to provide System Services
  - Inject necessary system services via Dependency Injection.
