+++ app/src/main/java/com/doyoonkim/knutice/di/AppModule.kt
@@ -1,6 +1,8 @@
 package com.doyoonkim.knutice.di
 
+import android.app.AlarmManager
 import android.app.Application
+import android.app.NotificationManager
 import android.content.Context
 import android.content.SharedPreferences
 import com.doyoonkim.common.di.ApplicationContext
@@ -21,5 +23,15 @@ object AppModule {
     @Singleton
     fun providesApplicationPreferences(@ApplicationContext context: Context): SharedPreferences =
         context.applicationContext.getSharedPreferences("app_pref", Context.MODE_PRIVATE)
+
+    @Provides
+    @Singleton
+    fun providesAlarmManager(@ApplicationContext context: Context) : AlarmManager =
+        context.getSystemService(Context.ALARM_SERVICE) as AlarmManager
+
+    @Provides
+    @Singleton
+    fun provideNotificationManager(@ApplicationContext context: Context): NotificationManager =
+        context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
 
 }
